Move connection handling to ConnectionManager
