+++ libs/cardreader/src/debug/java/com/woocommerce/android/cardreader/CardReaderManagerFactory.kt
@@ -3,13 +3,16 @@ package com.woocommerce.android.cardreader
 import com.woocommerce.android.cardreader.internal.CardReaderManagerImpl
 import com.woocommerce.android.cardreader.internal.payments.PaymentManager
 import com.woocommerce.android.cardreader.internal.TokenProvider
+import com.woocommerce.android.cardreader.internal.connection.ConnectionManager
 import com.woocommerce.android.cardreader.internal.payments.actions.CollectPaymentAction
 import com.woocommerce.android.cardreader.internal.payments.actions.CreatePaymentAction
 import com.woocommerce.android.cardreader.internal.payments.actions.ProcessPaymentAction
 import com.woocommerce.android.cardreader.internal.wrappers.LogWrapper
 import com.woocommerce.android.cardreader.internal.wrappers.PaymentIntentParametersFactory
 import com.woocommerce.android.cardreader.internal.wrappers.TerminalWrapper
+import kotlinx.coroutines.ExperimentalCoroutinesApi
 
+@ExperimentalCoroutinesApi
 object CardReaderManagerFactory {
     fun createCardReaderManager(cardReaderStore: CardReaderStore): CardReaderManager {
         val terminal = TerminalWrapper()
@@ -23,7 +26,8 @@ object CardReaderManagerFactory {
                 CreatePaymentAction(PaymentIntentParametersFactory(), terminal, logWrapper),
                 CollectPaymentAction(terminal, logWrapper),
                 ProcessPaymentAction(terminal, logWrapper)
-            )
+            ),
+            ConnectionManager(terminal, logWrapper)
         )
     }
 }
